Move app code into holochain-app crate
