+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "crates/mock_hdi",
   "crates/mr_bundle",
 
+  "crates/holochain_dna_types",
   "crates/holochain_integrity_types",
   "crates/holochain_zome_types",
   "crates/holochain_types",
@@ -28,6 +29,9 @@ members = [
   "crates/holochain_metrics",
   "crates/holochain_nonce",
   "crates/holochain_secure_primitive",
+
+  "crates/holochain-app",
+  # "crates/holochain-dna-bundle",
   "crates/holochain-services",
 
   "crates/hc",
