+++ crates/holochain-app/src/bundle.rs
@@ -2,21 +2,18 @@
 
 use std::{collections::HashMap, path::PathBuf, sync::Arc};
 
-use self::error::AppBundleResult;
-
-use super::{AppManifest, AppManifestValidated};
-use crate::prelude::*;
-
-#[allow(missing_docs)]
-mod error;
-pub use error::*;
+use crate::manifest::{AppManifestValidated, AppRoleManifestValidated};
+use crate::*;
 use futures::future::join_all;
+use holochain_dna_types::prelude::*;
 
-#[cfg(test)]
-mod tests;
+pub type AppBundleError = anyhow::Error;
+pub type AppBundleResult<T> = Result<T, AppBundleError>;
 
 /// A bundle of an AppManifest and collection of DNAs
-#[derive(Debug, Serialize, Deserialize, derive_more::From, shrinkwraprs::Shrinkwrap)]
+#[derive(
+    Debug, Serialize, Deserialize, derive_more::From, derive_more::Deref, derive_more::DerefMut,
+)]
 pub struct AppBundle(mr_bundle::Bundle<AppManifest>);
 
 impl AppBundle {
@@ -233,10 +230,9 @@ impl AppBundle {
                 };
             let expected_hash: DnaHash = hash.clone().into();
             if expected_hash != original_hash {
-                return Err(AppBundleError::CellResolutionFailure(
-                    role_name,
-                    format!("Hash mismatch: {} {}", expected_hash, original_hash),
-                ));
+                anyhow::bail!(
+                    "Could not resolve the app role '{role_name}' because of hash mismatch: {original_hash} != {expected_hash}"
+                );
             }
             dna_file
         } else {
@@ -259,16 +255,11 @@ impl AppBundle {
         modifiers: DnaModifiersOpt,
     ) -> AppBundleResult<(DnaFile, DnaHash)> {
         let bytes = self.resolve(location).await?;
-        let dna_bundle: DnaBundle = mr_bundle::Bundle::decode(&bytes)?.into();
+        let bundle = mr_bundle::Bundle::<ValidatedDnaManifest>::decode(&bytes)?;
+        let dna_bundle = DnaBundle::from(bundle);
         let (dna_file, original_hash) = dna_bundle.into_dna_file(modifiers).await?;
         Ok((dna_file, original_hash))
     }
-}
-
-/// This function is called in places where it will be necessary to rework that
-/// area after use_existing has been implemented
-pub fn we_must_remember_to_rework_cell_panic_handling_after_implementing_use_existing_cell_resolution(
-) {
 }
 
 /// The answer to the question:
@@ -311,6 +302,7 @@ impl AppRoleResolution {
 /// Specifies what step should be taken to provision a cell while installing an App
 #[warn(missing_docs)]
 #[derive(Debug)]
+#[allow(unused)]
 pub enum CellProvisioningOp {
     /// Create a new Cell from the given DNA file
     CreateFromDnaFile(DnaFile, u32),
@@ -328,3 +320,92 @@ pub enum CellProvisioningOp {
 /// Uninhabitable placeholder
 #[derive(Debug)]
 pub enum CellProvisioningConflict {}
+
+#[cfg(test)]
+mod tests {
+    use std::path::PathBuf;
+
+    use crate::bundle::AppRoleResolution;
+    use crate::manifest::app_manifest_v1::tests::{
+        app_manifest_fixture, app_manifest_properties_fixture,
+    };
+    use crate::manifest::DnaLocation;
+    use crate::*;
+    use ::fixt::prelude::*;
+    use holochain_dna_types::prelude::*;
+
+    use super::AppBundle;
+
+    async fn app_bundle_fixture(
+        modifiers: DnaModifiersOpt<YamlProperties>,
+    ) -> (AppBundle, DnaFile) {
+        let dna_wasm = DnaWasmHashed::from_content(DnaWasm::new_invalid()).await;
+        let fake_wasms = vec![dna_wasm.clone().into_content()];
+        let fake_zomes = vec![IntegrityZome::new(
+            "hi".into(),
+            ZomeDef::Wasm(WasmZome::new(dna_wasm.as_hash().clone())).into(),
+        )];
+        let dna_def_1 = DnaDef::unique_from_zomes(fake_zomes.clone(), vec![]);
+
+        let dna1 = DnaFile::new(dna_def_1, fake_wasms.clone()).await;
+
+        let path1 = PathBuf::from(format!("{}", dna1.dna_hash()));
+
+        let manifest = app_manifest_fixture(
+            Some(DnaLocation::Bundled(path1.clone())),
+            DnaHash::with_data_sync(dna1.dna_def()),
+            modifiers,
+        )
+        .await;
+
+        let resources = vec![(path1, DnaBundle::from_dna_file(dna1.clone()).unwrap())];
+
+        let bundle = AppBundle::new(manifest.into(), resources, PathBuf::from("."))
+            .await
+            .unwrap();
+        (bundle, dna1)
+    }
+
+    /// Test that an app with a single Created cell can be provisioned
+    #[tokio::test]
+    async fn provisioning_1_create() {
+        holochain_trace::test_run().ok();
+        let agent = fixt!(AgentPubKey);
+        let modifiers = DnaModifiersOpt {
+            properties: Some(app_manifest_properties_fixture()),
+            network_seed: Some("network_seed".into()),
+            origin_time: None,
+            quantum_time: None,
+        };
+        let (bundle, dna) = app_bundle_fixture(modifiers).await;
+
+        // Apply the modifier overrides specified in the manifest fixture
+        let dna = dna
+            .with_network_seed("network_seed".to_string())
+            .await
+            .with_properties(SerializedBytes::try_from(app_manifest_properties_fixture()).unwrap())
+            .await;
+
+        let cell_id = CellId::new(dna.dna_hash().to_owned(), agent.clone());
+
+        let resolution = bundle
+            .resolve_cells(
+                &std::collections::HashMap::new(),
+                agent.clone(),
+                Default::default(),
+            )
+            .await
+            .unwrap();
+
+        // Build the expected output.
+        // NB: this relies heavily on the particulars of the `app_manifest_fixture`
+        let role = AppRoleAssignment::new(cell_id, true, 50);
+
+        let expected = AppRoleResolution {
+            agent,
+            dnas_to_register: vec![(dna, None)],
+            role_assignments: vec![("role_name".into(), role)],
+        };
+        assert_eq!(resolution, expected);
+    }
+}
