+++ crates/holochain-app/src/lib.rs
@@ -1,0 +1,12 @@
+mod error;
+mod ported;
+mod state;
+
+mod bundle;
+pub mod manifest;
+
+pub use bundle::{AppBundle, AppBundleResult};
+pub use error::{AppError, AppResult};
+pub use manifest::*;
+
+pub use ported::*;
