+++ crates/holochain-app/src/manifest/current.rs
