+++ crates/holochain_dna_types/src/error.rs
