--- crates/holochain_types/src/dna.rs
@@ -1,22 +1,0 @@
-//! dna is a library for working with holochain dna files/entries.
-//!
-//! It includes utilities for representing dna structures in memory,
-//! as well as serializing and deserializing dna, mainly to json format.
-
-mod coordinator_bundle;
-mod dna_bundle;
-mod dna_file;
-mod dna_manifest;
-mod dna_store;
-mod error;
-mod ribosome_store;
-
-pub mod wasm;
-pub use coordinator_bundle::*;
-pub use dna_bundle::*;
-pub use dna_file::*;
-pub use dna_manifest::*;
-pub use dna_store::*;
-pub use error::*;
-pub use holo_hash::*;
-pub use ribosome_store::*;
