+++ crates/holochain_types/src/prelude.rs
@@ -1,9 +1,9 @@
 //! reexport some common things
 
+pub use holochain_app::*;
+pub use holochain_dna_types::prelude::*;
 pub use holochain_keystore::AgentPubKeyExt;
 pub use holochain_nonce::Nonce256Bits;
-pub use holochain_serialized_bytes::prelude::*;
-pub use holochain_zome_types::prelude::*;
 pub use std::convert::TryFrom;
 pub use std::convert::TryInto;
 
@@ -18,7 +18,6 @@ pub use crate::combinators::*;
 pub use crate::db::*;
 pub use crate::db_cache::*;
 pub use crate::dht_op::*;
-pub use crate::dna::wasm::*;
 pub use crate::dna::*;
 pub use crate::entry::*;
 pub use crate::link::*;
