+++ crates/holochain_types/src/web_app/web_app_bundle.rs
@@ -1,5 +1,6 @@
 use std::borrow::Cow;
 
+use holochain_app::AppBundle;
 use mr_bundle::{error::MrBundleResult, ResourceBytes};
 
 use super::WebAppManifest;
