+++ crates/kitsune_p2p/kitsune_p2p/Cargo.toml
@@ -19,7 +19,7 @@ derive_more = "0.99.11"
 futures = "0.3"
 ghost_actor = "=0.3.0-alpha.6"
 governor = "0.3.2"
-itertools = "0.11.0"
+itertools = "0.12"
 kitsune_p2p_fetch = { version = "^0.3.0-beta-dev.16", path = "../fetch" }
 kitsune_p2p_mdns = { version = "^0.3.0-beta-dev.1", path = "../mdns" }
 kitsune_p2p_proxy = { version = "^0.3.0-beta-dev.14", path = "../proxy" }
