+++ crates/release-automation/Cargo.toml
@@ -38,7 +38,7 @@ cargo-next = "0.1"
 toml_edit = "0.2"
 thiserror = "1"
 regex = "1.5"
-itertools = "0.10"
+itertools = "0.12"
 crates-index-helper = { git = "https://github.com/holochain/holochain-nixpkgs", rev = "d4701b21239be5223409123d7bdd06004af3b044" }
 
 # used for the example clippy fix-json
